% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricola.R
\name{agricola_get_metadata}
\alias{agricola_get_metadata}
\title{Returns dataframe containing metadata for items in Agricola catalog}
\usage{
agricola_get_metadata(res)
}
\arguments{
\item{res}{agricola query}
}
\value{
metadata dataframe
}
\description{
Returns dataframe containing metadata for items in Agricola catalog
}
