% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricola.R
\name{agricola_parse}
\alias{agricola_parse}
\title{Given agricola link return metadata dataframe}
\usage{
agricola_parse(agricola_link)
}
\arguments{
\item{agricola_link}{string}
}
\value{
metadata dataframe
}
\description{
Given agricola link return metadata dataframe
}
\examples{
res <- agricola_query(title = "Biological soil crusts: Ecology and management", author = "Jayne Belnap")
lapply(res$link, agricola_parse)

}
