% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agricola.R
\name{agricola_query}
\alias{agricola_query}
\title{Querying both NAL Agricola databases (journal articles and other materials) and parsing the search results.}
\usage{
agricola_query(
  rem_dr,
  title = NULL,
  author = NULL,
  keyword = NULL,
  isbn = NULL,
  language = NULL,
  pub = NULL,
  series = NULL,
  subject = NULL,
  source = NULL,
  agricola_search_url = "https://agricola.nal.usda.gov/vwebv/search"
)
}
\arguments{
\item{rem_dr}{remote driver}

\item{title}{string}

\item{author}{string}

\item{keyword}{string}

\item{isbn}{string}

\item{language}{string}

\item{pub}{string}

\item{series}{string}

\item{subject}{string}

\item{source}{string}

\item{agricola_search_url}{string optional argument}
}
\value{
dataframe of title and id links for top results (or all if fields blank).
}
\description{
Querying both NAL Agricola databases (journal articles and other materials) and parsing the search results.
}
\examples{

library(RSelenium)
rem_dr <- remoteDriver(port = 4445L)
rem_dr$open(silent = TRUE)

agricola_query(rem_dr, title = "Biological soil crusts: Ecology and management", author = "Jayne Belnap")

}
