% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm_parse}
\alias{blm_parse}
\title{Given a BLM ID string return a metadata dataframe row}
\usage{
blm_parse(row)
}
\arguments{
\item{string}{}
}
\value{

}
\description{
Given a BLM ID string return a metadata dataframe row
}
