% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm_query}
\alias{blm_query}
\title{Query BLM library with the given search parameters}
\usage{
blm_query(
  query = NULL,
  years = NULL,
  mediatypes = NULL,
  subjects = NULL,
  creators = NULL,
  languages = NULL
)
}
\arguments{
\item{query}{string, no vector}

\item{years}{string}

\item{mediatypes}{string}

\item{subjects}{string}

\item{creators}{string}

\item{languages}{string}
}
\value{
dataframe title, link
}
\description{
Query BLM library with the given search parameters
}
\examples{
blm_query(query = "wildlife", subjects = c("Wildlife conservation", "land"))
}
