% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_doi}
\alias{check_doi}
\title{Check to see if a string is potentially a valid DOI}
\usage{
check_doi(x)
}
\arguments{
\item{x}{string input to see if doi}
}
\value{
bool
}
\description{
Check to see if a string is potentially a valid DOI
}
