% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{create_combined}
\alias{create_combined}
\title{Create Combined column for authors and publishers}
\usage{
create_combined(authors, publishers)
}
\arguments{
\item{authors}{character vector}

\item{publisher}{character vector}
}
\description{
Create Combined column for authors and publishers
}
