% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{create_queries}
\alias{create_queries}
\title{Convert dataframe of citations into search queries}
\usage{
create_queries(citations)
}
\arguments{
\item{citations}{dataframe of citations}
}
\value{
search queries
}
\description{
Convert dataframe of citations into search queries
}
