% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_string}
\alias{empty_string}
\title{Check if the contents of a string should be considered empty}
\usage{
empty_string(x)
}
\arguments{
\item{x}{string to search, single value}
}
\value{
bool
}
\description{
Useful because of the ambiguity of our input data, use with purrr::discard
}
