% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_all}
\alias{index_all}
\title{Create collection containing nfsl, blm and treesearch datasets}
\usage{
index_all(
  nfsl_data = NULL,
  blm_data = NULL,
  treesearch_data = NULL,
  conn = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{conn}{optional solrclient connection object}

\item{overwrite}{optional bool delete and recreate collection if exists}

\item{nfsl}{optional dataframe of nfsl record metadata}

\item{blm}{optional dataframe of blm record metadata}

\item{treesearch}{optional dataframe of treesearch record metadata}
}
\description{
Create collection containing nfsl, blm and treesearch datasets
}
