% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_collection}
\alias{index_collection}
\title{Create collection for index}
\usage{
index_collection(collection)
}
\arguments{
\item{collection}{a dataframe used for indexing}
}
\description{
The collection passed in must have the following specified columns:
Title, Authors, Year, Publisher", Source, Misc, Journal Title, DOI
}
