% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_nfsl_data}
\alias{index_nfsl_data}
\title{Create index from nfsl dataset}
\usage{
index_nfsl_data(data = NULL, conn = NULL, overwrite = FALSE)
}
\arguments{
\item{data}{optional dataframe of nfsl data}

\item{conn}{optional solrclient connection object}

\item{overwrite}{optional bool delete and recreate collection if exists}
}
\description{
Create index from nfsl dataset
}
