% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_records}
\alias{index_records}
\title{Create collection for index}
\usage{
index_records(records, collection_name = paste0(substitute(records)))
}
\arguments{
\item{records}{a dataframe to be indexed}
}
\description{
The collection passed in must have the following specified columns:
Title, Authors, Year, Publisher, Source, Miscid, Journal Title, DOI
}
