% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_wos_data}
\alias{index_wos_data}
\title{Create index from Web of Science dataset}
\usage{
index_wos_data(data = NULL, conn = NULL, overwrite = FALSE)
}
\arguments{
\item{data}{optional dataframe of wos data}

\item{conn}{optional solrclient connection object}

\item{overwrite}{optional bool delete and recreate collection if exists}
}
\description{
Create index from Web of Science dataset
}
