% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{isValidColumns}
\alias{isValidColumns}
\title{Check to see if user has the number of matching columns specified}
\usage{
isValidColumns(df)
}
\arguments{
\item{df}{dataframe to validate if it has expected column names}
}
\description{
Check to see if user has the number of matching columns specified
}
