% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_authors}
\alias{list_authors}
\title{change author format from anystyle output to simple set
Anystyle makes json-style outputs that split out first/middle/surname for each author}
\usage{
list_authors(x)
}
\arguments{
\item{x}{author string set that follows anystyle format}
}
\value{
string
}
\description{
change author format from anystyle output to simple set
Anystyle makes json-style outputs that split out first/middle/surname for each author
}
