% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsl.R
\name{nfsl_get_metadata_all}
\alias{nfsl_get_metadata_all}
\title{Returns dataframe containing metadata for all items in NFSL catalog}
\usage{
nfsl_get_metadata_all()
}
\description{
Returns dataframe containing metadata for all items in NFSL catalog
}
