% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsl.R
\name{nfsl_parse}
\alias{nfsl_parse}
\title{Given NFSL ID return metadata dataframe}
\usage{
nfsl_parse(
  nfsl_id,
  nfsl_doc_url = "https://nfsl.contentdm.oclc.org/digital/api/collections/"
)
}
\arguments{
\item{nfsl_id}{string nfsl id link}

\item{nfsl_doc_url}{string optional argument}
}
\value{
metadata dataframe
}
\description{
Given NFSL ID return metadata dataframe
}
\examples{
nfsl_parse("p17053coll2/id/310050")
}
