% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsl.R
\name{nfsl_query}
\alias{nfsl_query}
\title{Querying the National Forest Service Library and parsing the search results.}
\usage{
nfsl_query(
  title = NULL,
  creators = NULL,
  publisher = NULL,
  contributors = NULL,
  date = NULL,
  identifier = NULL,
  source = NULL
)
}
\arguments{
\item{title}{string}

\item{creators}{string}

\item{publisher}{string}

\item{contributors}{string}

\item{date}{string}

\item{identifier}{string}

\item{source}{string}
}
\value{
dataframe of title and id links for top results (or all if fields blank).
}
\description{
Querying the National Forest Service Library and parsing the search results.
}
\examples{
nfsl_query(creators = c("arno", "tomback"))
                 
}
