% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_string}
\alias{normalize_string}
\title{Normalize a string}
\usage{
normalize_string(x)
}
\arguments{
\item{x}{string to normalize}
}
\value{
normalized string
}
\description{
Normalizes a string by setting to lowercase, filtering out punct and digits
}
