% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anystyle.R
\name{normalize_years}
\alias{normalize_years}
\title{Normalize anystyle dates}
\usage{
normalize_years(years)
}
\arguments{
\item{years}{character or numeric vector}
}
\value{
years normalized
}
\description{
Normalize anystyle dates
}
