% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anystyle.R
\name{parse_anystyle}
\alias{parse_anystyle}
\title{Convert anystyle data to the input format expected for searches}
\usage{
parse_anystyle(anystyle_df)
}
\arguments{
\item{anystyle_entry}{dataframe}
}
\value{
dataframe
}
\description{
Convert anystyle data to the input format expected for searches
}
