% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{preprocess_anystyle_entry}
\alias{preprocess_anystyle_entry}
\title{Preprocess an anystyle entry so that it can be searched with}
\usage{
preprocess_anystyle_entry(anystyle_entry)
}
\arguments{
\item{anystyle_entry}{dataframe}
}
\value{
dataframe
}
\description{
Preprocess an anystyle entry so that it can be searched with
}
