% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{preprocess_blm_data}
\alias{preprocess_blm_data}
\title{Preprocess BLM data}
\usage{
preprocess_blm_data(blm_data)
}
\arguments{
\item{blm_data}{dataframe of blm record metadata}
}
\description{
Preprocess BLM data
}
