% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess a collection}
\usage{
preprocess_data(collection)
}
\arguments{
\item{collection}{a dataframe used for indexing}
}
\value{
tibble of the processed data
}
\description{
Preprocess a collection
}
