% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{preprocess_nfsl_data}
\alias{preprocess_nfsl_data}
\title{Preprocess NFSL data}
\usage{
preprocess_nfsl_data(nfsl_data)
}
\arguments{
\item{nfsl_data}{dataframe of nfsl record metadata}
}
\value{
dataframe of preprocessed ready to be indexed data
}
\description{
Preprocess NFSL data
}
