% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{preprocess_treesearch_data}
\alias{preprocess_treesearch_data}
\title{Preprocess Treesearch data}
\usage{
preprocess_treesearch_data(treesearch_data)
}
\arguments{
\item{treesearch_data}{dataframe of nfsl record metadata}
}
\description{
Preprocess Treesearch data
}
