% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{preprocess_wos_data}
\alias{preprocess_wos_data}
\title{Preprocess Web of Science data}
\usage{
preprocess_wos_data(wos_data)
}
\arguments{
\item{wos_data}{dataframe of web of science data}
}
\value{
dataframe of preprocessed ready to be indexed data
}
\description{
Preprocess Web of Science data
}
