% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rdetect}
\alias{rdetect}
\title{Helper function to detect if two sets of string vectors EQUAL to each other}
\usage{
rdetect(string, pattern)
}
\arguments{
\item{string}{a set of strings}

\item{pattern}{the format to search against the string}
}
\value{
a list of bool
}
\description{
Helper function to detect if two sets of string vectors EQUAL to each other
}
