% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_collection}
\alias{search_collection}
\title{Given search query, get top matches}
\usage{
search_collection(query_string, collection_name, topn = 1, conn = NULL)
}
\arguments{
\item{query_string}{created with construct_query function}

\item{collection_name}{string to search against}

\item{topn}{optional number or results to return}

\item{conn}{optional SolrClient object}
}
\value{
dataframe of top 5 results
}
\description{
Given search query, get top matches
}
