% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_construct_query}
\alias{search_construct_query}
\title{Convert Anystyle item to a search query string
takes row, returns string}
\usage{
search_construct_query(anystyle_entry)
}
\arguments{
\item{anystyle_entry}{single row from anystyle output}
}
\value{
string suitable for solr search
}
\description{
Convert Anystyle item to a search query string
takes row, returns string
}
