% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesearch.R
\name{treesearch_get_metadata_all}
\alias{treesearch_get_metadata_all}
\title{Returns dataframe containing metadata for all items in treesearch catalog}
\usage{
treesearch_get_metadata_all()
}
\description{
Returns dataframe containing metadata for all items in treesearch catalog
}
