% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesearch.R
\name{treesearch_get_metadata_one}
\alias{treesearch_get_metadata_one}
\title{Given Treesearch ID return metadata dataframe}
\usage{
treesearch_get_metadata_one(
  treesearch_id,
 
    treesearch_doc_url = "https://data.fs.usda.gov/research/westernstubservices/productsservicesentry?"
)
}
\arguments{
\item{treesearch_id}{treesearch id number}

\item{treesearch_doc_url}{string optional argument}
}
\description{
Given Treesearch ID return metadata dataframe
}
\examples{
treesearch_get_metadata_one(28951)
}
