% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesearch.R
\name{treesearch_web_query}
\alias{treesearch_web_query}
\title{Perform a search on Treesearch's online interface}
\usage{
treesearch_web_query(
  keywords = "",
  authorlastname = "",
  yearfrom = "",
  yearto = "",
  station = "",
  series = "",
  volume = "",
  treesearch_url = "https://www.fs.usda.gov/treesearch/search?"
)
}
\arguments{
\item{keywords}{string}

\item{authorlastname}{string}

\item{yearfrom}{string}

\item{yearto}{string}

\item{station}{string}

\item{series}{string}

\item{volume}{string}

\item{treesearch_url}{string optional argument}
}
\value{
dataframe title,id,text for top 20 results (or all if fields blank).
}
\description{
Perform a search on Treesearch's online interface
}
\examples{
treesearch_web_query(keywords = "increment-borer methods fire history", 
                 authorlastname <- "Arno")
}
