% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Check to see if user has the number of matching columns specified}
\usage{
validate_columns(cols, expected)
}
\arguments{
\item{cols}{char colnames to check}

\item{expected}{char names to check against}

\item{return}{bool}
}
\description{
Check to see if user has the number of matching columns specified
}
