% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_years}
\alias{validate_years}
\title{Check if years column contains only numbers between 1800 and 2025}
\usage{
validate_years(x)
}
\arguments{
\item{x}{years column}
}
\value{
bool
}
\description{
Check if years column contains only numbers between 1800 and 2025
}
