% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_blm_data}
\alias{index_blm_data}
\title{Create index from blm dataset}
\usage{
index_blm_data(data = NULL, conn = NULL)
}
\arguments{
\item{data}{optional dataframe of nfsl data}

\item{conn}{optional solrclient connection object}
}
\description{
Create index from blm dataset
}
