% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_treesearch_data}
\alias{index_treesearch_data}
\title{Create index from Treesearch dataset}
\usage{
index_treesearch_data(data = NULL, conn = NULL)
}
\arguments{
\item{data}{optional dataframe of treesearch data}

\item{conn}{optional solrclient connection object}
}
\description{
Create index from Treesearch dataset
}
