% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_collection}
\alias{search_collection}
\title{Given Anystyle row, get top matches}
\usage{
search_collection(collection_name, query_string, conn = NULL)
}
\arguments{
\item{collection_name}{one of blm,treesearch,nfsl}

\item{query_string}{created with construct_query function}

\item{conn}{optional SolrClient object}
}
\value{
dataframe of top 5 results
}
\description{
Given Anystyle row, get top matches
}
